restructured the source code tree
